# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIComp Electronics Comparison Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filtering and pagination
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: category
          in: query
          schema: { type: string }
        - name: brand
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, popular, price_asc, price_desc, rating, trending] }
        - name: minPrice
          in: query
          schema: { type: integer }
        - name: maxPrice
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /v1/products/trending:
    get:
      operationId: getTrendingProducts
      tags: [products]
      summary: Get trending products
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Trending products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /v1/products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured and popular products for homepage
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomepageData"
  /v1/products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
  /v1/products/{slug}/prices:
    get:
      operationId: getProductPrices
      tags: [products]
      summary: Get all prices for a product
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Price list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceList"
  /v1/products/{slug}/similar:
    get:
      operationId: listSimilarProducts
      tags: [products]
      summary: Get similar products
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Similar products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"

  # ─── CATEGORIES ─────────────────────────────────────────────────────────────
...
  /v1/search:
    get:
      operationId: searchProducts
      tags: [search]
      summary: Full-text product search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
  /v1/search/autocomplete:
    get:
      operationId: searchAutocomplete
      tags: [search]
      summary: Autocomplete suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Autocomplete results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutocompleteResults"

  # ─── COMPARISONS ────────────────────────────────────────────────────────────
...
  /v1/deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List current deals and price drops
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Deals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealList"
  /v1/deals/flash:
    get:
      operationId: getFlashDeals
      tags: [deals]
      summary: Get flash/hot deals for homepage
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 8 }
      responses:
        "200":
          description: Flash deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealList"

  # ─── LAUNCHES ───────────────────────────────────────────────────────────────
...
  /v1/guides:
    get:
      operationId: listGuides
      tags: [content]
      summary: List buying guides and content
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: category
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Guide list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuideList"
  /v1/guides/{slug}:
    get:
      operationId: getGuide
      tags: [content]
      summary: Get guide by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Guide detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guide"

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ProductSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        thumbnail: { type: ["string", "null"] }
        brand: { type: string }
        brandSlug: { type: string }
        category: { type: string }
        categorySlug: { type: string }
        categoryEmoji: { type: ["string", "null"] }
        lowestPrice: { type: ["integer", "null"] }
        mrp: { type: ["integer", "null"] }
        discount: { type: ["integer", "null"] }
        avgRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        overallScore: { type: ["integer", "null"] }
        isTrending: { type: boolean }
        isFeatured: { type: boolean }
        status: { type: string }
        releaseDate: { type: ["string", "null"] }
        viewCount: { type: integer }
        aiPros: { type: array, items: { type: string } }
        aiCons: { type: array, items: { type: string } }
      required: [id, name, slug, brand, brandSlug, category, categorySlug, reviewCount, isTrending, isFeatured, status, viewCount, aiPros, aiCons]
    ProductSpec:
      type: object
      properties:
        groupName: { type: string }
        specName: { type: string }
        specValue: { type: string }
        unit: { type: ["string", "null"] }
        sortOrder: { type: integer }
      required: [groupName, specName, specValue, sortOrder]
    ProductPrice:
      type: object
      properties:
        id: { type: string }
        seller: { type: string }
        sellerLogo: { type: ["string", "null"] }
        price: { type: integer }
        mrp: { type: ["integer", "null"] }
        discount: { type: ["integer", "null"] }
        inStock: { type: boolean }
        affiliateUrl: { type: ["string", "null"] }
        cashback: { type: ["string", "null"] }
        deliveryDays: { type: ["integer", "null"] }
        lastChecked: { type: string }
      required: [id, seller, price, inStock, lastChecked]
    ProductDetail:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        thumbnail: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        brand: { type: string }
        brandSlug: { type: string }
        category: { type: string }
        categorySlug: { type: string }
        categoryEmoji: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        shortDesc: { type: ["string", "null"] }
        aiSummary: { type: ["string", "null"] }
        aiPros: { type: array, items: { type: string } }
        aiCons: { type: array, items: { type: string } }
        avgRating: { type: number }
        reviewCount: { type: integer }
        viewCount: { type: integer }
        compareCount: { type: integer }
        isTrending: { type: boolean }
        isFeatured: { type: boolean }
        releaseDate: { type: ["string", "null"] }
        gameScore: { type: ["integer", "null"] }
        batteryScore: { type: ["integer", "null"] }
        cameraScore: { type: ["integer", "null"] }
        perfScore: { type: ["integer", "null"] }
        coolingScore: { type: ["integer", "null"] }
        valueScore: { type: ["integer", "null"] }
        overallScore: { type: ["integer", "null"] }
        specs: { type: array, items: { $ref: "#/components/schemas/ProductSpec" } }
        prices: { type: array, items: { $ref: "#/components/schemas/ProductPrice" } }
        seoTitle: { type: ["string", "null"] }
        seoDesc: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
      required: [id, name, slug, brand, brandSlug, category, categorySlug, avgRating, reviewCount, viewCount, compareCount, isTrending, isFeatured, aiPros, aiCons, specs, prices, tags]
    ProductList:
      type: object
      properties:
        products: { type: array, items: { $ref: "#/components/schemas/ProductSummary" } }
        total: { type: integer }
        page: { type: integer }
        pages: { type: integer }
      required: [products, total, page, pages]
    HomepageData:
      type: object
      properties:
        trending: { type: array, items: { $ref: "#/components/schemas/ProductSummary" } }
        featured: { type: array, items: { $ref: "#/components/schemas/ProductSummary" } }
        recentLaunches: { type: array, items: { $ref: "#/components/schemas/LaunchItem" } }
        topDeals: { type: array, items: { $ref: "#/components/schemas/DealItem" } }
        popularComparisons: { type: array, items: { $ref: "#/components/schemas/ComparisonPair" } }
        stats: { $ref: "#/components/schemas/SiteStats" }
      required: [trending, featured, recentLaunches, topDeals, popularComparisons, stats]
    SiteStats:
      type: object
      properties:
        totalProducts: { type: integer }
        totalSellers: { type: integer }
        pricesUpdated: { type: string }
        totalComparisons: { type: integer }
      required: [totalProducts, totalSellers, pricesUpdated, totalComparisons]
...
    PriceList:
      type: object
      properties:
        prices: { type: array, items: { $ref: "#/components/schemas/ProductPrice" } }
        lowestPrice: { type: ["integer", "null"] }
        highestPrice: { type: ["integer", "null"] }
        lastUpdated: { type: string }
      required: [prices, lastUpdated]

    # ─── CATEGORY ─────────────────────────────────────────────────────────────
...
    SearchResult:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        image: { type: ["string", "null"] }
        brand: { type: string }
        category: { type: string }
        price: { type: ["string", "null"] }
        score: { type: ["number", "null"] }
      required: [id, name, slug, brand, category]
    SearchResults:
      type: object
      properties:
        results: { type: array, items: { $ref: "#/components/schemas/SearchResult" } }
        total: { type: integer }
        query: { type: string }
        page: { type: integer }
        pages: { type: integer }
      required: [results, total, query, page, pages]
    AutocompleteResults:
      type: object
      properties:
        results: { type: array, items: { $ref: "#/components/schemas/SearchResult" } }
      required: [results]

    # ─── COMPARISONS ──────────────────────────────────────────────────────────
...
    ComparisonPair:
      type: object
      properties:
        slug: { type: string }
        productA: { type: string }
        productB: { type: string }
        productASlug: { type: string }
        productBSlug: { type: string }
        viewCount: { type: integer }
      required: [slug, productA, productB, productASlug, productBSlug, viewCount]
...
    DealItem:
      type: object
      properties:
        productId: { type: string }
        productName: { type: string }
        productSlug: { type: string }
        thumbnail: { type: ["string", "null"] }
        brand: { type: string }
        category: { type: string }
        currentPrice: { type: integer }
        mrp: { type: integer }
        discount: { type: integer }
        seller: { type: string }
        affiliateUrl: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
      required: [productId, productName, productSlug, brand, category, currentPrice, mrp, discount, seller]
    DealList:
      type: object
      properties:
        deals: { type: array, items: { $ref: "#/components/schemas/DealItem" } }
        total: { type: integer }
      required: [deals, total]

    # ─── LAUNCHES ─────────────────────────────────────────────────────────────
    LaunchItem:
      type: object
      properties:
        id: { type: string }
        productName: { type: string }
        productSlug: { type: ["string", "null"] }
        brand: { type: string }
        category: { type: string }
        categoryEmoji: { type: ["string", "null"] }
        thumbnail: { type: ["string", "null"] }
        expectedDate: { type: ["string", "null"] }
        confirmedDate: { type: ["string", "null"] }
        expectedPrice: { type: ["integer", "null"] }
        confirmedPrice: { type: ["integer", "null"] }
        isConfirmed: { type: boolean }
        status: { type: string }
      required: [id, productName, brand, category, isConfirmed, status]
...
    Guide:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: ["string", "null"] }
        body: { type: string }
        type: { type: string }
        category: { type: ["string", "null"] }
        viewCount: { type: integer }
        publishedAt: { type: ["string", "null"] }
        seoTitle: { type: ["string", "null"] }
        seoDesc: { type: ["string", "null"] }
        ogImage: { type: ["string", "null"] }
        isAiGen: { type: boolean }
      required: [id, title, slug, body, type, viewCount, isAiGen]
    GuideList:
      type: object
      properties:
        guides: { type: array, items: { $ref: "#/components/schemas/Guide" } }
        total: { type: integer }
        page: { type: integer }
        pages: { type: integer }
      required: [guides, total, page, pages]

    # ─── ADMIN ────────────────────────────────────────────────────────────────
...