# lib/api-zod/src/generated/types/guideList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIComp Electronics Comparison Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Guide } from './guide';

export interface GuideList {
  guides: Guide[];
  total: number;
  page: number;
  pages: number;
}