# lib/api-zod/src/generated/types/homepageData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIComp Electronics Comparison Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ComparisonPair } from './comparisonPair';
import type { DealItem } from './dealItem';
import type { LaunchItem } from './launchItem';
import type { ProductSummary } from './productSummary';
import type { SiteStats } from './siteStats';

export interface HomepageData {
  trending: ProductSummary[];
  featured: ProductSummary[];
  recentLaunches: LaunchItem[];
  topDeals: DealItem[];
  popularComparisons: ComparisonPair[];
  stats: SiteStats;
}